+++ module/research/AutoClassifierProvider/src/AutoClassifierProvider.h
@@ -35,14 +35,14 @@ namespace research {
 
     private:
         int ballEdgeBuffer  = 0;
-        int ballLightnessMin = 0;
-        int ballLightnessMax = 255;
+        uint8_t ballLightnessMin = 0;
+        uint8_t ballLightnessMax = 255;
         int goalEdgeBuffer  = 0;
-        int goalLightnessMin = 0;
-        int goalLightnessMax = 255;
+        uint8_t goalLightnessMin = 0;
+        uint8_t goalLightnessMax = 255;
         int fieldEdgeBuffer = 0;
-        int fieldLightnessMin = 0;
-        int fieldLightnessMax = 255;
+        uint8_t fieldLightnessMin = 0;
+        uint8_t fieldLightnessMax = 255;
         int lineEdgeBuffer  = 0;
         int lineLightnessMin = 0;
         int lineLightnessMax = 255;
